requirements.txt
pandas 
mlflow
notebook==6.5.6
numpy
scikit-learn
matplotlib
seaborn
jupyter-server==1.24.0
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
flask
flask-cors
pandas
python-dotenv
ensure
ipykernel